Put TestAssignmentModel to separate file and inherit from AssignmentModel
