+++ Scripts/assignment.py
@@ -2,45 +2,8 @@ import os
 import logging
 import inro.emme.desktop.app as _app
 import inro.modeller as _m
-import omx
-import numpy
 import parameters as param
-
-class TestAssignmentModel:
-    def __init__(self, matrix_dir):
-        self.path = matrix_dir
-    
-    def assign(self, time_period, matrices):
-        """Get travel impedance matrices for one time period from files."""
-        mtxs = {}
-        mtxs["time"] = self.get_matrices(time_period, "time")
-        mtxs["cost"] = self.get_matrices(time_period, "cost")
-        mtxs["dist"] = self.get_matrices(time_period, "dist")
-        return mtxs
-    
-    def get_matrices(self, time_period, mtx_type):
-        file_name = os.path.join(self.path, mtx_type+'_'+time_period+".omx")
-        costs_file = omx.openFile(file_name)
-        matrices = dict.fromkeys(param.emme_mtx[mtx_type].keys())
-        for mtx in matrices:
-            matrices[mtx] = numpy.array(costs_file[mtx])
-        costs_file.close()
-        return matrices
-    
-    def get_zone_numbers(self):
-        file_name = os.path.join(self.path, "time_aht.omx")
-        costs_file = omx.openFile(file_name)
-        # zone_numbers = costs_file.mapentries("zone_number")
-        zone_numbers = [5, 6, 7]
-        costs_file.close()
-        return zone_numbers
-    
-    def get_mapping(self):
-        file_name = os.path.join(self.path, "time_aht.omx")
-        costs_file = omx.openFile(file_name)
-        mapping = costs_file.mapping("zone_number")
-        costs_file.close()
-        return mapping
+from assignment_model import AssignmentModel
 
 class EmmeAssignmentModel:
     def __init__(self, filepath):
@@ -590,6 +553,8 @@ class EmmeAssignmentModel:
         self.logger.info("Transit assignment performed for scenario " 
                         + str(scen_id))
         
+AssignmentModel.register(EmmeAssignmentModel)
+
 class Car:
     def __init__(self, ass_class, value_of_time_inv, 
                  od_travel_times=None, link_costs="@rumsi"):
